+++ R/charlatan-package.R
@@ -8,12 +8,13 @@
 #' @name charlatan-package
 #' @importFrom R6 R6Class
 #' @importFrom whisker whisker.render
-#' @importFrom tibble as_data_frame
+#' @importFrom tibble as_tibble
 #' @aliases charlatan
 #' @docType package
 #' @keywords package
 #' @author Scott Chamberlain \email{myrmecocystus+r@@gmail.com}
 #' @author Kyle Voytovich
+#' @author Martin Pedersen
 #'
 #' @section Package API:
 #'