+++ R/misc-provider.R
@@ -1,27 +1,20 @@
-#' MiscProvider
-#'
+#' @title MiscProvider
+#' @description miscellaneous methods
 #' @export
 #' @keywords internal
-#' @details
-#' **Methods**
-#' 
-#' - `language_locale_codes` - list of language locales supported by 
-#'     Linux Mint
-#' - `locale()` - locale
-#' - `language_code()` - language code
-#'
-#' @format NULL
-#' @usage NULL
 #' @examples
 #' (x <- MiscProvider$new())
 #' x$language_locale_codes
 #' x$language_code()
 #' x$locale()
+#' x$boolean()
+#' x$null_boolean()
 MiscProvider <- R6::R6Class(
   inherit = BaseProvider,
   'MiscProvider',
   public = list(
     # Locales supported by Linux Mint from `/usr/share/i18n/SUPPORTED`
+    #' @field language_locale_codes (list) locale codes by locale family
     language_locale_codes = list(
         'aa' = c('DJ', 'ER', 'ET'),
         'af' = c('ZA'),
@@ -210,16 +203,18 @@ MiscProvider <- R6::R6Class(
         'zh' = c('CN', 'HK', 'SG', 'TW'),
         'zu' = c('ZA')
     ),
+    
+    #' @description get a random boolean, `TRUE` or `FALSE`
+    #' @param chance_of_getting_true (integer) an integer, default: 50
+    boolean = function(chance_of_getting_true = 50) {
+      super$random_int(1, 100) <= chance_of_getting_true
+    },
 
-    # def boolean(self, chance_of_getting_true=50):
-    #     return self.generator.random.randint(1, 100) <= chance_of_getting_true
-
-    # def null_boolean(self):
-    #     return {
-    #         0: None,
-    #         1: True,
-    #         -1: False
-    #     }[self.generator.random.randint(-1, 1)]
+    #' @description get a random boolean, `TRUE` or `FALSE`, or NULL
+    null_boolean = function() {
+      take <- as.character(super$random_int(-1, 1))
+      list(`0` = NULL, `1` = TRUE, `-1` = FALSE)[[take]]
+    },
 
     # def binary(self, length=(1 * 1024 * 1024)):
     #     """ Returns random binary blob.
@@ -229,103 +224,15 @@ MiscProvider <- R6::R6Class(
     #     blob = [self.generator.random.randrange(256) for o in range(length)]
     #     return bytes(blob) if sys.version_info[0] >= 3 else bytearray(blob)
 
-    # def md5(self, raw_output=False):
-    #     """
-    #     Calculates the md5 hash of a given string
-    #     :example 'cfcd208495d565ef66e7dff9f98764da'
-    #     """
-    #     res = hashlib.md5(str(self.generator.random.random()).encode('utf-8'))
-    #     if raw_output:
-    #         return res.digest()
-    #     return res.hexdigest()
-
-    # def sha1(self, raw_output=False):
-    #     """
-    #     Calculates the sha1 hash of a given string
-    #     :example 'b5d86317c2a144cd04d0d7c03b2b02666fafadf2'
-    #     """
-    #     res = hashlib.sha1(str(self.generator.random.random()).encode('utf-8'))
-    #     if raw_output:
-    #         return res.digest()
-    #     return res.hexdigest()
-
-    # def sha256(self, raw_output=False):
-    #     """
-    #     Calculates the sha256 hash of a given string
-    #     :example '85086017559ccc40638fcde2fecaf295e0de7ca51b7517b6aebeaaf75b4d4654'
-    #     """
-    #     res = hashlib.sha256(
-    #         str(self.generator.random.random()).encode('utf-8'))
-    #     if raw_output:
-    #         return res.digest()
-    #     return res.hexdigest()
-
+    #' @description get a random locale
     locale = function() {
       lc = self$language_code()
       paste0(lc, '_', super$random_element(self$language_locale_codes[[lc]]))
     },
 
+    #' @description random language code
     language_code = function() {
       super$random_element(names(self$language_locale_codes))
     }
-
-    # def uuid4(self):
-    #     """
-    #     Generates a random UUID4 string.
-    #     """
-    #     # Based on http://stackoverflow.com/q/41186818
-    #     return str(uuid.UUID(int=self.generator.random.getrandbits(128)))
-
-    # def password(
-    #         self,
-    #         length=10,
-    #         special_chars=True,
-    #         digits=True,
-    #         upper_case=True,
-    #         lower_case=True):
-    #     """
-    #     Generates a random password.
-    #     @param length: Integer. Length of a password
-    #     @param special_chars: Boolean. Whether to use special characters !@#$%^&*()_+
-    #     @param digits: Boolean. Whether to use digits
-    #     @param upper_case: Boolean. Whether to use upper letters
-    #     @param lower_case: Boolean. Whether to use lower letters
-    #     @return: String. Random password
-    #     """
-    #     choices = ""
-    #     required_tokens = []
-    #     if special_chars:
-    #         required_tokens.append(
-    #             self.generator.random.choice("!@#$%^&*()_+"))
-    #         choices += "!@#$%^&*()_+"
-    #     if digits:
-    #         required_tokens.append(self.generator.random.choice(string.digits))
-    #         choices += string.digits
-    #     if upper_case:
-    #         required_tokens.append(
-    #             self.generator.random.choice(string.ascii_uppercase))
-    #         choices += string.ascii_uppercase
-    #     if lower_case:
-    #         required_tokens.append(
-    #             self.generator.random.choice(string.ascii_lowercase))
-    #         choices += string.ascii_lowercase
-
-    #     assert len(
-    #         required_tokens) <= length, "Required length is shorter than required characters"
-
-    #     # Generate a first version of the password
-    #     chars = [self.generator.random.choice(choices) for x in range(length)]
-
-    #     # Pick some unique locations
-    #     random_indexes = set()
-    #     while len(random_indexes) < len(required_tokens):
-    #         random_indexes.add(
-    #             self.generator.random.randint(0, len(chars) - 1))
-
-    #     # Replace them with the required characters
-    #     for i, index in enumerate(random_indexes):
-    #         chars[index] = required_tokens[i]
-
-    #     return ''.join(chars)
-    )
+  )
 )